requirements.txt
fastapi
pandas
uvicorn
python-dotenv
pydantic-settings